requirements.txt
azure_storage==0.37.0
Flask==3.0.3
psycopg2==2.9.10
